+++ crates/coordinator/dances/tests/shared_test/test_stage_new_holon.rs
@@ -11,7 +11,7 @@ use hdk::prelude::*;
 use holochain::sweettest::*;
 use holochain::sweettest::{SweetCell, SweetConductor};
 use rstest::*;
-use dances::holon_dances::{build_get_all_holons_dance_request, build_stage_new_holon_dance_request};
+use dances::holon_dance_adapter::{build_get_all_holons_dance_request, build_stage_new_holon_dance_request};
 use dances::dance_response::{DanceResponse, ResponseStatusCode};
 use dances::dance_response::ResponseBody::{Holons, Index};
